Start refactoring heatmap/contour into reusable parts
